pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "lyzr-pgm"
version = "0.1.0"
description = "Lyzr Agent Manager scripts for creating, managing, and running Manager/Role agents with YAML workflows."
authors = [
    { name = "Lyzr AI Team" }
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pyyaml",
    "httpx",
    "pytz",
    "tzlocal"
]

[project.scripts]
run-list = "scripts.run_list_iterate:main"
create-agent = "scripts.create_agent:main"
create-manager = "scripts.create_manager:main"
create-manager-from-yaml = "scripts.create_manager_from_yaml:main"
run-agent = "scripts.run_agent:main"
